Rename to WmtsCapabilities
