+++ BruTile/GlobalSuppressions.cs
@@ -5,5 +5,5 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-[assembly: SuppressMessage("Trimming", "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code", Justification = "<Pending>", Scope = "member", Target = "~M:BruTile.Wmts.WmtsParser.Parse(System.IO.Stream,BruTile.Wmts.BoundingBoxAxisOrderInterpretation)~System.Collections.Generic.List{BruTile.Web.HttpTileSource}")]
+[assembly: SuppressMessage("Trimming", "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code", Justification = "<Pending>", Scope = "member", Target = "~M:BruTile.Wmts.WmtsCapabilitiesParser.Parse(System.IO.Stream,BruTile.Wmts.BoundingBoxAxisOrderInterpretation)~System.Collections.Generic.List{BruTile.Web.HttpTileSource}")]
 [assembly: SuppressMessage("Trimming", "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code", Justification = "<Pending>", Scope = "member", Target = "~M:BruTile.Tms.TileMapParser.CreateTileSource(System.IO.Stream,System.String,System.Collections.Generic.Dictionary{System.String,System.String},BruTile.Cache.IPersistentCache{System.Byte[]},System.Func{System.Uri,System.Threading.CancellationToken,System.Threading.Tasks.Task{System.Byte[]}})~BruTile.TileSource")]
